First model (D/D/1/K-1) half done.

+++ src/main/java/models/QueueSystemInput.java
@@ -6,11 +6,11 @@ import io.reactivex.annotations.Nullable;
 public class QueueSystemInput {
     private Double arrivalRate;
     private Double serviceRate;
-    private Double numberOfServers;
-    private Double systemCapacity;
+    private int numberOfServers;
+    private int systemCapacity;
 
     public QueueSystemInput(@NonNull Double arrivalRate, @NonNull Double serviceRate,
-                            @Nullable Double numberOfServers, @Nullable Double systemCapacity) {
+                            @Nullable int numberOfServers, @Nullable int systemCapacity) {
         this.arrivalRate = arrivalRate;
         this.serviceRate = serviceRate;
         this.numberOfServers = numberOfServers;
@@ -25,11 +25,11 @@ public class QueueSystemInput {
         return serviceRate;
     }
 
-    public Double getNumberOfServers() {
+    public int getNumberOfServers() {
         return numberOfServers;
     }
 
-    public Double getSystemCapacity() {
+    public int getSystemCapacity() {
         return systemCapacity;
     }
